% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/break_even.R
\docType{data}
\name{break_even}
\alias{break_even}
\title{Detailed run times for analysis of break-even point}
\format{\preformatted{'data.frame':	1806000 obs. of  5 variables:
 $ prob: chr  "uniform" "uniform" "uniform" "uniform" ...
 $ expr: Factor w/ 5 levels "ccrank","crank",..: 1 1 1 1 1 1 1 1 1 1 ...
 $ time: num  42920 7587 5337 5096 4697 ...
 $ r   : num  0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 ...
 $ n   : int  10 10 10 10 10 10 10 10 10 10 ...
}}
\usage{
break_even
}
\description{
Run times measured on an Intel(R) Xeon(R) CPU E5-2643 clocked at 3.30 GHz
with 10 MB cache.
}
\keyword{datasets}

