% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/timings.R
\docType{data}
\name{timings}
\alias{timings}
\title{Run times}
\format{\preformatted{'data.frame':	48000 obs. of  5 variables:
 $ prob: chr  "uniform" "uniform" "uniform" "uniform" ...
 $ expr: Factor w/ 5 levels "ccrank","crank",..: 1 1 1 1 1 1 1 1 1 1 ...
 $ time: num  64138 9311 7269 6545 6343 ...
 $ r   : num  0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 ...
 $ n   : int  31 31 31 31 31 31 31 31 31 31 ...
}}
\usage{
timings
}
\description{
Run times measured on an Intel(R) Xeon(R) CPU E5-2643 clocked at 3.30 GHz
with 10 MB cache.
}
\keyword{datasets}

