% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timings.R, R/break_even.R
\name{timings}
\alias{break_even}
\alias{timings}
\title{Run time data}
\format{A data frame with 5 columns:

\describe{
  \item{\code{prob}}{
    A description of the probability distribution used.
    See \code{data_raw/benchmark.R} for details.
  }
  \item{\code{expr}}{
    Function name without the \code{sample_int_} prefix.
  }
  \item{\code{time}}{
    Run time in nanoseconds, as measured by
    \code{\link[microbenchmark]{microbenchmark}}
  }
  \item{\code{r}}{
    Ratio between the \code{size} and \code{n} arguments.
  }
  \item{\code{n}}{
    The \code{n} argument.
  }
}}
\description{
Run times measured on an Intel(R) Xeon(R) CPU E5-2643 clocked at 3.30 GHz
with 10 MB cache, running Debian Squeeze, R 3.1.3 and gcc 4.4.5-8.
The data are created by the corresponding scripts in the \code{data_raw}
directory.
}
\details{
\code{timings} contains run times for a larger range of values
  for the \code{n} argument.

\code{break_even} contains detailed run times for the analysis of
break-even points between the various implementations.
}

