% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timings.R
\docType{data}
\name{timings}
\alias{timings}
\alias{timings_sort}
\alias{break_even}
\title{Run time data}
\format{
An object of class \code{data.frame} with 25200 rows and 5 columns.

A data frame with 5 columns:

\describe{
\item{\code{prob}}{
A description of the probability distribution used.
See \code{data_raw/benchmark.R} for details.
}
\item{\code{expr}}{
Function name without the \code{sample_int_} prefix.
}
\item{\code{time}}{
Run time in nanoseconds, as measured by
\code{\link[microbenchmark:microbenchmark]{microbenchmark::microbenchmark()}}
}
\item{\code{r}}{
Ratio between the \code{size} and \code{n} arguments.
}
\item{\code{n}}{
The \code{n} argument.
}
}
}
\usage{
timings_sort
}
\description{
Run times measured on an Intel(R) Xeon(R) CPU X5680 clocked at 3.33 GHz
with 12 MB cache, running RedHat Enterprise Linux, R 3.2.3 and gcc 4.8.5,
using version 0.4 of the \code{wrswoR} package.
The data are created by the corresponding scripts in the \code{data_raw}
directory.
}
\details{
\code{timings} contains run times for a larger range of values
for the \code{n} argument.

\code{timings_sort} contains run times for sorting probabilities
with the given distributions.

\code{break_even} contains detailed run times for the analysis of
break-even points between the various implementations.
}
\examples{
head(timings)
head(break_even)
}
\keyword{datasets}
